requirements.txt
absl-py==1.4.0
addict==2.4.0
ansi2html==1.8.0
appdirs==1.4.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bracex==2.3.post1
cached-property==1.5.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
cycler==0.11.0
Cython==0.29.28
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.9
einops==0.6.1
fastjsonschema==2.18.1
filelock==3.12.2
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fsspec==2023.1.0
gdown==4.7.1
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.23.0
google-auth-oauthlib==0.4.6
grpcio==1.58.0
h5py==3.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources==5.12.0
ipython==7.34.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_core==4.12.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdx-truly-sane-lists==1.3
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-material==9.2.7
mkdocs-material-extensions==1.1.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multimethod==1.7
natsort==8.4.0
nbformat==5.5.0
nest-asyncio==1.5.8
ninja==1.10.2.3
numba @ file:///croot/numba_1670258325998/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
oauthlib==3.2.2
open3d==0.16.0
OpenEXR==1.3.9
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
paginate==0.5.6
pandas==1.3.5
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.18.0
pointnet2-cuda==0.0.0
prompt-toolkit==3.0.39
protobuf==3.19.4
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pymdown-extensions==10.2.1
pyparsing==3.1.1
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyvista==0.33.2
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
safetensors==0.3.3
scikit-learn==1.0.2
scipy==1.7.3
scooby==0.7.2
sentry-sdk==1.30.0
setproctitle==1.3.2
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.4.1
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.10.1
torch-scatter==2.1.1
torchmetrics @ file:///home/conda/feedstock_root/build_artifacts/torchmetrics_1678720682204/work
torchvision==0.11.2
tqdm==4.62.3
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.16
vtk==9.2.6
wandb==0.15.10
watchdog==3.0.0
wcmatch==8.4.1
wcwidth==0.2.8
Werkzeug==2.2.3
widgetsnbextension==4.0.9
zipp @ file:///croot/zipp_1672387121353/work
